refactor(mobile): move the Home page to a split directory
