+++ mobile/lib/app/features/home/home.dart
@@ -1,0 +1,1 @@
+export 'view/home_page.view.dart';
